Refactored Template specific code into separate class
